Feature/temp

+++ DuLich/DAL/TourDal.cs
@@ -1,6 +1,8 @@
 ﻿using DuLich.BUS;
 using DuLich.Entity;
+using System;
 using System.Collections.Generic;
+using System.Data.Entity;
 using System.Data.Entity.Migrations;
 using System.Linq;
 using System.Threading.Tasks;
@@ -16,19 +18,36 @@ namespace DuLich.DAL
         {
             return context.Touris.ToList();
         }
-        public static Task AddOrUpdate(Tour newTour)
+        public static Task AddOrUpdate(Tour tour)
         {
-            context.Touris.AddOrUpdate(newTour);
-            return context.SaveChangesAsync();
+            var result = context.Touris.SingleOrDefault(b => b.MaTour == tour.MaTour);
+            if (result != null)
+            {
+                result.ChiTietTour = tour.ChiTietTour;
+                result.DoanTour = tour.DoanTour;
+                result.GiaTour = tour.GiaTour;
+                result.LoaiTour = tour.LoaiTour;
+                result.TenTour = tour.TenTour;
+                return context.SaveChangesAsync();
+            }
+            else
+            {
+                context.Touris.Add(tour);
+                return context.SaveChangesAsync();
+            }
         }
         public static Task Delete(Tour touris)
         {
-            context.Touris.Remove(touris);
+            var result = context.Touris.SingleOrDefault(b => b.MaTour == touris.MaTour);
+            if (result != null)
+            {
+                context.Touris.Remove(touris);
+            }
             return context.SaveChangesAsync();
         }
         public static List<Doan> GetListDoanByTourId(int id)
         {
-            return context.Doans.Where(c => c.Touris.Id == id).ToList();
+            return context.Doans.Where(c => c.Tour.MaTour == id).ToList();
         }
 
     
@@ -36,11 +55,11 @@ namespace DuLich.DAL
         public static List<DiaDiem> GetListDiaDiemByTour(Tour tour)
         {
             List<DiaDiem> danhSachDiaDiemCuaTor = new List<DiaDiem>();
-            if (tour.ChiTietTours == null)
+            if (tour.ChiTietTour == null)
                 return danhSachDiaDiemCuaTor;
-            foreach (ChiTietTour chiTietTour in tour.ChiTietTours.OrderBy(c => c.ThuTu).ToList())
+            foreach (ChiTietTour chiTietTour in tour.ChiTietTour.OrderBy(c => c.ThuTu).ToList())
             {
-                DiaDiem diaDiem = context.DiaDiem.Where(c => c.MaDienDiem == chiTietTour.diaDiem.MaDienDiem).First();
+                DiaDiem diaDiem = context.DiaDiem.Where(c => c.MaDienDiem == chiTietTour.DiaDiem.MaDienDiem).First();
                 if (!danhSachDiaDiemCuaTor.Contains(diaDiem))
                     danhSachDiaDiemCuaTor.Add(diaDiem);
             }
@@ -49,9 +68,9 @@ namespace DuLich.DAL
 
         public static void DeleteAllChiTietByTour(Tour tour)
         {
-            if (tour.ChiTietTours != null)
+            if (tour.ChiTietTour != null)
             {
-                context.ChiTietTour.RemoveRange(tour.ChiTietTours);
+                context.ChiTietTour.RemoveRange(tour.ChiTietTour);
             }
         }
 
@@ -66,7 +85,9 @@ namespace DuLich.DAL
 
         public static List<Tour> GetListTourHasPrice()
         {
-            return GetAll().Where(c => c.Gias!=null && c.Gias.Count >0).ToList();
+            return GetAll().Where(c => c.GiaTour!=null && c.GiaTour.Count >0).ToList();
         }
+
+       
     }
 }
